fix 1. 디렉토리 저장 경로 app.yml에 설정 후 참조하는 방식으로 변경 2. 엔티티 매우 많이 수정 1) 클라이언트로 부터 받은 json 변수명, 타입과 서버 엔티티 일치화 2) 받은 num 값을 id로 찾아 detail 엔티티에 참조 (productList, Member)
add : 받은 데이터들 DB에 저장하는 기능

+++ src/main/java/com/example/repairproductprogram/controller/SearchProductNum.java
@@ -30,7 +30,7 @@ public class SearchProductNum {
         }
     }
     @GetMapping("/exists")
-    public ResponseEntity<Boolean> checkProductExists(@RequestParam("productNum") int productNum) {
+    public ResponseEntity<Boolean> checkProductExists(@RequestParam("productNum") Long productNum) {
         boolean exists = productListRepository.existsByProductNum(productNum);
         return ResponseEntity.ok(exists);
     }
@@ -45,7 +45,7 @@ public class SearchProductNum {
 
         if ("product_num".equals(filterBy)) {
             try {
-                Integer product_num = Integer.parseInt(searchTerm);
+                Long product_num = Long.parseLong(searchTerm);
                 return productListRepository.findByProductNum(product_num);
             } catch (NumberFormatException e) {
                 return Collections.emptyList();
